Bugfix: Handle f_directory_not status as an invalid directory.
The status f_directory_not is returned with the path exists but is not a directory.

The fake program can fail when doing something like this:
  fake clean make -w /arbitrary_non_directory_file +V

The error message returned is not very helpful:
  ERROR: (674) when calling function f_file_stat().

This is a known file-related error and so it is now handled appropraitely.
The new error message looks like:
  ERROR: Invalid directory while trying to access directory '/arbitrary_non_directory_file'.