Progress: Add simple runtime testing scripts and testfile for fss_read.
This is an incredibly basic implementation.
I didn't bother to make is as flexible and portable as it could otherwise be.

Provide a way to test the runtime results of the fss_read programs.
I only got so far as to add tests for FSS Basic (FSS-0000) and FSS Extended (FSS-0001).

The tests for the FSS Extended (FSS-0001) currently has some failures.
I've noticed that I have not even implemented the --select functionality.
This is the likely cause for several of the test failures.

For now, I just cheated and used the 0.6.x version of the project to generate the expected output.
Any bugs in that output will be matched (ideally there are no bugs).
I did do a quick glance at some of these and they do look fine.

The source test files are copies of files I have used in the past but never made publicly available until now.