Regression: Incorrect parameter location used, breaking parameter processing.
The regression is introduced in commit 4c0ca86048e431de2025c39a3f5078b7248192eb.

I completely forgot that "location" is for the parameter location and is not the last value location.
Some part of me wanted a last value location.
Revert back to using "values.used - 1" instead of "location".