Bugfix: Quotes in FSS read functions are not being properly handled.
At some point in time the intended behavior changed from using the literal quote character to an enumeration representing the quote character on FSS read calls.
This transition is incomplete and broken.

Finish the transition, passing the enumeration representation rather than the literal character.

Remove an unnecessary initialization on the quote because the private function being called also initializes the variable.

Relocate the f_uint8s_increase() call to be after changing the error status to ensure that the quotes array is properly incremented.

Update the documentation comments to better clarify this behavior in regards to the enumeration.