Progress: Fix issues in runtime tests for fss_read not relating to Embedded List and Payload.
More runtime tests now pass.

The Extended List tests still fail in some cases.
I noticed that the Extended List line count is not handling the closing brace.
The count is inaccurate.
This is a bug in the 0.6.x branch and needs to be fixed there.
The fix is in place for this branch but the tests currently fail because they generated expect files are incorrect in this case.

There are other similar problems in different FSS read programs regarding --total and parameters like --at or --line that need to be fixed in the 0.6.x branch as well.

The Embedded List and Payload need special handling and are not addressed here.

Other things regarding fss_read not done:
- The help printing needs to be updated for each program.
- Any remaining todo and fixme.
- The pipe input and output.

There is likely a lot of room for improvement but I am going to ignore such things for now as my priorities are elsewhere.