Bugfix: FSS Extended List Read program should print new line after Object.
In the case where Object is printed but not Content a new line must be printed.
This is done to be consistent with the other programs, such as FSS Basic List Read.

The 0.7.x and later versions should probably implement something like -L/--line_end to perform this behavior.
In which case the default behavior would be what the code was doing here prior to this commit.