Feature: Add f_socket_name_host() and f_socket_name_peer().
Add missing network socket related functions for getting the "name" of the local or remote connections.

The local name is retuned as a string.

The remote name is returned on the f_socket_t structure.
I considered doing the work necessary to better generalize this but I decided not to.
I can leave that for once I am better experience in the network programming.