Cleanup: Do not include man page documentation if there are no man pages.
This never generated an error locally because the directories happen to exist.
Empty directories are generally not stored in git and so when unit tests were run on Github the missing directory problem becomes exposed.

This is not a bug because the related bug is fixed in commit ecddd4d16504d3e2d9502d6a4dfe67658dce98dc.
This only silences the newly added warning.