Update: The f_socket enumerations and comments, the f_socket_t structure, the f_socket build settings, and the f_socket unit tests.
The _di_f_socket_protocol_e_ enumerations are missing the trailing "_e" in the name.

As stated in a previous commit (90d0f47999fd5615f2f3550d688f63d49e590482), update the documentation comments for the f_socket_protocol_family_e enumerations.

Change the ordering of properties in the f_socket_t to have length property after the address property.

Set the default string for the name property to be f_string_static_t_initialize instead of f_string_empty_s.

Disable the braces warning in f_console.
The brace usage here is correct, and so the warning is disabled.

Fix the unit tests as they were not updated after changes in the previous commit (90d0f47999fd5615f2f3550d688f63d49e590482) and in this commit.