Update: Specifications, adding explicit version date and change license line format.
Add a version, which contains the date of the last change to each of the specifications.
There will eventually be a specification version number such as "1.2.3".

The license line in the comments at the top of each specification is of this form: "license: ".
Remove the colon so that the comment, if extracted, could be processed as something simple like FSS-0000.