Cleanup: Documentation comments regarding structures.
Add "Properties: " and related styling to the documentation comments for type definition structures.

Fix capitalization and other formatting issues that I noticed.

I did not focus on any of the programs.
However, some program related files happened to be open and so I updated those out of convenience.
A follow up commit will be needed to address the programs and any other files in which I may have missed in this pass.