Progress: Add memory array allocation functionality and begin to use it where possible.
The pattern for handling the standard array structure is now well established.
Provide a memory array functionality for handling the most basic form.

This should reduce repeated code where possible.
It would be interesting if this reduces the size of the resulting binary but with inline optimizations it may be possible that this actually increases the size of the resulting binary.

Update the comments and tests.

There are a lot of files that will need to be updated to use these new functions.
Most of these are in the f_type_array project.