Progress: Continue memory array related changes, replacing the f_memory_structure entirely.
The f_memory_array set of functions now fully replaces the f_memory_structure functions and related macros.
I think those structure macros are nifty but they are not in use enough anymore and the f_memory_array structure is what I am going forward with.

There is still an enormouse amount of changes to do.