Update: Finish memory array related changes.
There are still the append() and append_all() function to change but I am going to leave that for another time.
This finishes the recent set of Progress commits addressing the memory array related code.

I have observed a notable reduction in binary size but the primary goal here is reduction of duplicate code and simplification of code.

I should eventually be able to re-introduce macros to map these helper functions back to the base array type to reduce binary size by a significant amount.
This should only be done once the API has stabilized more.