Progress: Begin simplifying memory design, starting with f_number_unsigned_t array structures.
I find that the use of custom per-type functions to simplify the usability of the library.
I also find that this increases the complexity and size of the code base.

I believe that over time this will become more complex and even harder to maintain.
I have decided to start changing the design to go back to simple allocation methods that just require more variables.
This simplifies the code and in theory will make it faster by having fewer function calls on the stack.
This may even reduce the resulting binary size (and I hope it does).

The down side is that the users of the library will once more have longer function call lines.

This commit is the beginning of the redesign process.

I also added the "append" and "append all" methods for f_memory_array calls.