Progress: Continue re-designing of the memory logic.
This wraps up the simple memory append all functions with added unit tests.
I anticipate that I need to do the more complex ones as well.

This adds the adjust and resize callbacks for use in more complex memory functions.
The more complex memory functions for adjust and resize are implemented with unit tests.

The unit tests for the type array callbacks are incomplete because I did not get time to finish writing them.
Those unit tests are written but are incomplete and do not yet work.