Cleanup: Change structure of fll_execute_program() checks.
This is a follow up to the commit 1b43ceefc20836137ed95e19d275dda44d8a678d.

The previous commit is passing NULL to strrchr().
This should be fine, but does produce a warning.

Clean up the code and play it safe by changing the design to only call strrchr() when the value passed would be non-NULL.