Bugfix: Stop using f_memory_array_append_all() for an array of IKI data.
There is some logic mistake, structural mistake, design mistake, or an oversight.
Just do not use it for now and instead use a traditional for loop.

I may get rid of f_memory_array_append_all() in the future or I may spend time investigating what my mistake is in this use case.