Regression: Incorrect dependencies breaking some builds, specifically the stand alone builds.
I'm not sure when this happened but at a some point I cleaned up some of the dependencies but failed to perform the appropraite checks with the stand alone builds.
There are stale files, dependencies, and references that are causing the stand alone builds to fail.

Remove or add the correct dependencies as appropriate.