Feature: Add f_string_range_double_t, updating build settings and unit tests appropriately.
I need a string range structure that has two ranges for some given string for the network related code.

The unit tests are updated or added as appropriate.
Any bugs discovered in the related unit tests are fixed.

The build data is rebuilt.

I originally wrote f_string_range_t to be specific to strings.
I now think that this needs to be conceptually more generalized.
The purpose of such a generalization is to simplify the code logic and design.
Renaming f_string_range_t to f_range_t and now renaming f_string_range_double_t into f_range_double_t is the intended change.
This would then be moved out of the f_string project and into the f_type project.
This is too much to do in the scope of this commit.

Instead, the f_string_range_double_t is being added under f_string_range_double_t.
I will then follow up this commit with a refactor of the f_string_range_t and f_string_range_double_t.