Update: Further re-design f_socket_connect().
This is a follow up to commit 596997daebd53246b205d0221c5c616bfafb7d23.

Make sure to set the appropriate sizeof for the socket length.

Update the documentation comments to communicate that the socket length is updated on the function calls.

Make f_socket_connect()'s socket parameter a pointer.
Set the socket length in f_socket_connect() to the appropriate size.

Add documentation comments regarding the pointers that must not be NULL in f_socket.

Update the related unit tests.