Refactor: The backtick into the grave, for the specifications.
This is a terminology change in the specifications and the code (both API or ABI) does not need to be changed.
Changing the code would be ideal, but this would break API and likely ABI.

The 0.6 version series will continue to use the word "backtick" in the code itself for this reason.