Progress: Continue working on fl_fss_payload_header_map().
Mostly focus on private_fl_payload_header_map_maps().
There is a lot of similar code design with subtle but key differences.
This is going to be really easy to make an mistake with.
Writing unit tests is going to be the ideal way to ensure the logic here is correct.

I did a lot of back and forth jumping around while trying to remember where I was last and what I need to do.

Make the build settings files more consistent in regards to threads and coverage modes.