Progress: Continue refactoring firewall from 0.6.x to 0.7.x/0.8.x.
This is very close to be ready for review and testing.
I converted most of the buffers into the new structure but I need to check for isolation.
I am pretty sure that there are currently some isolation problems with the FSS Object and Content ranges.

I also noticed some typos and such that will need to be fixed (this therefore probably does not compile just yet).