Cleanup: Switch from _ui_ format string to _un_ format string to match the print format arguments.
The _un_ format string is provided to directly match '%un' which is intended to directly match f_number_unsigned_t.
This is intended to make things easier (and safer) to hack.
That is, changing the data size of f_number_unsigned_t.

The _ul_ format string is specifically intended to be an unsigned long, period.