Update: Finish refactoring firewall from 0.6.x to 0.7.x/0.8.x.
I have tested this out and compared the show results between the 0.6.x version and the 0.7.x version to confirm consistent results.
From this point forward problems are to be up for consideration as possible regressions.

Separate out more of the functions into separate files.

I reviewed the header includes and removed the headers that do not have apparent direct use.