Update: Documentation man pages.
This repeats the process outlined in commit 63be746d22d66870ae89f9575c82db7d0b784e32.

The following changes are made:
- TAB_SIZE = 2
- PROJECT_NUMBER = 0.6.9

Setting "JAVADOC_AUTOBRIEF = yes" may be something worth looking into but for now the default of "JAVADOC_AUTOBRIEF = no" is being used.

It is unclear whether or not the "@see" is being processed without looking closes (and given the size of this, I did not do that).
I decided to leave the SEE ALSO sections that are generated, regardless of their state.

I noticed and removed man page documentation comments that somehow ended up in unit test data files.

The program function documentation is also generated and provided now.
This was not done so in the past.

Performing the process of generating and then reviewing the generated man pages exposed a number of problems that I fixed in commits prior to this commit.
Most of the changes are code clean ups but a security concern regarding handling a NULL pointer in a parameter was caught and identified during this process.

The commit referenced above mentioned the doxy2man but never documented the process.
The doxy2man used can be found at https://github.com/gsauthof/doxy2man.git .
The commit hash used for the generation is 5ce113f4d2a3fc6712f8eb8606a6b0899dc6f8d1 (dated Wed Aug 31 09:06:46 2016 +0200).

The doxy2man tool was performed against every XML file generated from doxywizard via a Bash for loop over each XML file.

I may make a copy of this doxy2man tool to help preserve this useful tool.

This is a massive set of changes (almost 1600 man pages) and so my review is rather lax and generalized.